removed unwanted ModView code
